refactor: Aggregate server and ensured make test pass
