+++ Makefile
@@ -7,13 +7,12 @@ fmt:
 
 .PHONY: dbmateup
 dbmateup:
-	cd internal/cfp  && go run github.com/amacneil/dbmate/v2 up
-	cd internal/dkui  && go run github.com/amacneil/dbmate/v2 up
+	cd internal/infrastructure  && go run github.com/amacneil/dbmate/v2 up
 
 .PHONY: vet
 vet: dbmateup
 	go vet ./...
-	go run github.com/sqlc-dev/sqlc/cmd/sqlc vet -f internal/sqlc.yaml
+	go run github.com/sqlc-dev/sqlc/cmd/sqlc vet -f internal/infrastructure/db/sqlc.yaml
 
 .PHONY: test
 test: fmt vet
