+++ internal/infrastructure/.env
@@ -1,0 +1,1 @@
+DATABASE_URL="mysql://user:password@127.0.0.1:13306/dkui"
